+++ lib/flutter_aws_chime.dart
@@ -1,5 +1,8 @@
-
 import 'flutter_aws_chime_platform_interface.dart';
+export 'models/meeting_event.model.dart';
+export 'models/meeting.model.dart';
+export 'models/message.model.dart';
+export 'models/video_tile.model.dart';
 
 class FlutterAwsChime {
   Future<String?> getPlatformVersion() {